feat(content): add i18n context api system
Add SSR-safe i18n system using Svelte Context API:
- setI18nContext(): initialize context in layout
- useI18n(): access reactive translations in components
- hasI18nContext(): check if context is available

Benefits over module-level state:
- SSR-safe: state doesn't persist between server requests
- testability: easy to isolate tests with different locales
- nested components: local locale override possible
- Svelte 5 best practices: recommended pattern for shared data

Closes #45
