+++ src/lib/components/common/ConnectivityBanner.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { ConnectivityBanner as BaseConnectivityBanner } from '$lib/ui';
-	import { i18n } from '$lib/content';
+	import { useI18n } from '$lib/content';
 
 	/**
 	 * ConnectivityBanner - Pre-configured connectivity banner with i18n labels.
@@ -19,6 +19,7 @@
 
 	let { error }: Props = $props();
 
+	const i18n = useI18n();
 	const ui = $derived(i18n.ui);
 
 	const content = $derived({
