+++ src/lib/components/common/PageHeader.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import type { Snippet } from 'svelte';
 	import { PageHeader as BasePageHeader } from '$lib/ui';
-	import { i18n } from '$lib/content';
+	import { useI18n } from '$lib/content';
 
 	/**
 	 * PageHeader - Pre-configured page header with i18n labels.
@@ -48,6 +48,7 @@
 		class: className = ''
 	}: Props = $props();
 
+	const i18n = useI18n();
 	const a11y = $derived(i18n.accessibility);
 </script>
