+++ src/lib/components/dashboard/ArticlesCountCard.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { ArticlesCountCard as BaseArticlesCountCard } from '$lib/ui';
-	import { i18n } from '$lib/content';
+	import { useI18n } from '$lib/content';
 	import type { TInstitution } from '$lib/server/openalex';
 
 	/**
@@ -22,6 +22,7 @@
 
 	let { selectedOrganizations, hasConsent = false }: Props = $props();
 
+	const i18n = useI18n();
 	const ui = $derived(i18n.ui);
 
 	const content = $derived({
