+++ src/lib/components/dashboard/ProfileCard.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { ProfileCard as BaseProfileCard } from '$lib/ui';
-	import { i18n } from '$lib/content';
+	import { useI18n } from '$lib/content';
 
 	/**
 	 * ProfileCard - Pre-configured profile card with i18n labels.
@@ -23,6 +23,7 @@
 
 	let { email, isComplete = false, class: className = '' }: Props = $props();
 
+	const i18n = useI18n();
 	const ui = $derived(i18n.ui);
 
 	const content = $derived({
