+++ src/lib/components/dashboard/ThemePreferencesCard.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { ThemePreferencesCard as BaseThemePreferencesCard } from '$lib/ui';
-	import { i18n } from '$lib/content';
+	import { useI18n } from '$lib/content';
 
 	/**
 	 * ThemePreferencesCard - Pre-configured theme preferences card with i18n labels.
@@ -49,6 +49,7 @@
 		class: className = ''
 	}: Props = $props();
 
+	const i18n = useI18n();
 	const ui = $derived(i18n.ui);
 
 	const content = $derived({
