+++ src/lib/components/login/LoginForm.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { LoginForm as BaseLoginForm } from '$lib/ui';
-	import { i18n } from '$lib/content';
+	import { useI18n } from '$lib/content';
 	import type { AuthContent } from '$lib/content/types';
 
 	/**
@@ -27,6 +27,7 @@
 
 	let { form, class: className = '' }: Props = $props();
 
+	const i18n = useI18n();
 	const auth = $derived(i18n.auth);
 	const a11y = $derived(i18n.accessibility);
